final version

+++ DataGenerator/src/main/java/com/jswiente/phd/prototype/DataGenerator/Distribution.java
@@ -2,6 +2,6 @@ package com.jswiente.phd.prototype.DataGenerator;
 
 public interface Distribution {
 	
-	public long getInterval(double lambda);
+	public double getInterval(double lambda);
 
 }
